chandamama_agentic_v2/requirements.txt
streamlit>=1.32.0
openai>=1.12.0
qdrant-client>=1.6.0
sentence-transformers>=3.0.0
python-dotenv>=1.0.1
tqdm>=4.66.0
numpy
transformers>=4.42.0

huggingface_hub>=0.20.0
# accelerate
# protobuf
# tiktoken
# einops

groq>=0.4.0
groq
